+++ src/tests.rs
@@ -3,6 +3,38 @@
 // licensed under GPL-3.0-or-later
 
 mod parse {
+    mod frac {
+        use crate::parse::{parse_frac, ParseFracErr};
+        use core::num::NonZeroU8;
+
+        #[test]
+        fn basic() {
+            assert_eq!(parse_frac("1", NonZeroU8::new(1).unwrap()), Ok(1));
+            assert_eq!(parse_frac("2", NonZeroU8::new(1).unwrap()), Ok(2));
+            assert_eq!(
+                parse_frac("23", NonZeroU8::new(1).unwrap()),
+                Err(ParseFracErr::ExcessDigits { idx: 1 })
+            );
+            assert_eq!(parse_frac("23", NonZeroU8::new(2).unwrap()), Ok(23));
+            assert_eq!(parse_frac("2", NonZeroU8::new(2).unwrap()), Ok(20));
+            assert_eq!(
+                parse_frac("24🪴21", NonZeroU8::new(5).unwrap()),
+                Err(ParseFracErr::ParseDigit {
+                    idx: 2,
+                    len: 4,
+                    err: "g".parse::<u8>().unwrap_err()
+                })
+            );
+            {
+                let s = (u64::from(u32::MAX) + 1).to_string();
+                assert_eq!(
+                    parse_frac(&s, NonZeroU8::new(s.len() as _).unwrap()),
+                    Err(ParseFracErr::NumeratorOverflow { idx: s.len() - 1 })
+                );
+            }
+        }
+    }
+
     mod unit {
         // TODO: test unit format
     }